requirements.txt
flask
numpy 
tensorflow
torch
torchvision
pillow
